optimizer: split optimizer.go from cmd/ctr-remote/commands/optimize.go
